Move dialogs to own folder
