+++ nft-studio/src/components/Dialogs/CreateCollectionDialog.tsx
@@ -13,16 +13,16 @@ import {
 import React, { useEffect, useState } from 'react'
 import { useQueryClient } from 'react-query'
 import { Redirect } from 'react-router'
-import { ButtonGroup } from '../components/ButtonGroup'
-import { useWeb3 } from '../components/Web3Provider'
-import { collectionMetadataSchema } from '../schemas'
-import { createCollectionMetadata } from '../utils/createCollectionMetadata'
-import { getFileDataURI } from '../utils/getFileDataURI'
-import { useAsyncCallback } from '../utils/useAsyncCallback'
-import { useBalance } from '../utils/useBalance'
-import { useCentrifugeTransaction } from '../utils/useCentrifugeTransaction'
-import { fetchMetadata } from '../utils/useMetadata'
-import { useCentrifuge } from './CentrifugeProvider'
+import { collectionMetadataSchema } from '../../schemas'
+import { createCollectionMetadata } from '../../utils/createCollectionMetadata'
+import { getFileDataURI } from '../../utils/getFileDataURI'
+import { useAsyncCallback } from '../../utils/useAsyncCallback'
+import { useBalance } from '../../utils/useBalance'
+import { useCentrifugeTransaction } from '../../utils/useCentrifugeTransaction'
+import { fetchMetadata } from '../../utils/useMetadata'
+import { ButtonGroup } from '../ButtonGroup'
+import { useCentrifuge } from '../CentrifugeProvider'
+import { useWeb3 } from '../Web3Provider'
 
 // TODO: replace with better fee estimate
 const CREATE_FEE_ESTIMATE = 2
