+++ nft-studio/src/pages/InvestmentsToken.tsx
@@ -6,7 +6,7 @@ import { ButtonGroup } from '../components/ButtonGroup'
 import { CardHeader } from '../components/CardHeader'
 import { useCentrifuge } from '../components/CentrifugeProvider'
 import { ConnectButton } from '../components/ConnectButton'
-import { InvestRedeemDialog } from '../components/InvestRedeemDialog'
+import { InvestRedeemDialog } from '../components/Dialogs/InvestRedeemDialog'
 import { LabelValueList } from '../components/LabelValueList'
 import { PageHeader } from '../components/PageHeader'
 import { PageWithSideBar } from '../components/PageWithSideBar'
