+++ nft-studio/src/pages/NFT.tsx
@@ -3,16 +3,16 @@ import * as React from 'react'
 import { NavLink, useParams } from 'react-router-dom'
 import { BuyDialog } from '../components/BuyDialog'
 import { useCentrifuge } from '../components/CentrifugeProvider'
+import { RemoveListingDialog } from '../components/Dialogs/RemoveListingDialog'
+import { SellDialog } from '../components/Dialogs/SellDialog'
+import { TransferDialog } from '../components/Dialogs/TransferDialog'
 import { Identity } from '../components/Identity'
 import { PageHeader } from '../components/PageHeader'
 import { PageWithSideBar } from '../components/PageWithSideBar'
 import { AnchorPillButton } from '../components/PillButton'
-import { RemoveListingDialog } from '../components/RemoveListingDialog'
 import { RouterLinkButton } from '../components/RouterLinkButton'
-import { SellDialog } from '../components/SellDialog'
 import { SplitView } from '../components/SplitView'
 import { TextWithPlaceholder } from '../components/TextWithPlaceholder'
-import { TransferDialog } from '../components/TransferDialog'
 import { nftMetadataSchema } from '../schemas'
 import { parseMetadataUrl } from '../utils/parseMetadataUrl'
 import { useAddress } from '../utils/useAddress'
